Moved TNT library to utilities
