+++ prody/dynamics/bbenmtools.c
@@ -1,6 +1,6 @@
 /*****************************************************************************/
 /*                                                                           */
-/*                   Tools for RTB calculations in ProDy.                    */
+/*                   Tools for bbENM calculations in ProDy.                    */
 /*                                                                           */
 /*****************************************************************************/
 /* Author: Cihan Kaya, She Zhang */
