+++ prody/utilities/tnt/tnt_array1d_utils.h
